add get_property API

+++ xconfig/xpredefined_configurations.h
@@ -12,10 +12,10 @@
 #include <cstdint>
 #include <limits>
 
-constexpr uint64_t TOP_UNIT = 1e6; // 1TOP = 1e6 uTOP
-constexpr uint64_t TOTAL_ISSUANCE = 200 * 1e8 * TOP_UNIT;
-#define ASSET_TOP(num) ((uint64_t)((num)*TOP_UNIT))
-#define ASSET_uTOP(num) ((uint64_t)(num))
+constexpr uint64_t TOP_UNIT = static_cast<uint64_t >(1e6); // 1TOP = 1e6 uTOP
+constexpr uint64_t TOTAL_ISSUANCE = static_cast<uint64_t>(200 * 1e8 * TOP_UNIT);
+#define ASSET_TOP(num) (static_cast<uint64_t>((num)*TOP_UNIT))
+#define ASSET_uTOP(num) (static_cast<uint64_t>(num))
 #define TOP_UNIT_LENGTH 6
 #define TOP_MAX_LENGTH 12 // enough for 20,000,000,000 tokens
